Handle wildcard "+" segments
A relatively common way to allow negotation to multiple serialization
structures for a single media type is to use "+" notation within the
mediatype:

- `application/vnd.book+json`: JSON serialization of
  `application/vnd.book` mediatype.
- `application/vnd.book+xml`: XML serialization of
  `application/vnd.book` mediatype.

In particular, this is useful for determining what _general_
serialization is used for an incoming request in order to select an
appropriate deserializer.

With the current 2.X branch, I expected the following to work:

```php
use Negotiation\Negotiator;

$mediaType = (new Negotiator)->getBest($accept, [
    'application/json',
    'application/*+json',
    'application/xml',
    'application/*+xml',
]);
```

However, it failed on each of the cases listed above.

This patch provides additional negotiation routines within
`Negotiation\Negotiator::match` in order to allow such negotiation to
work as expected.